@@ -1,21 +1,21 @@
 module Lti2Tp
   class Registration < ActiveRecord::Base
 
-    END_REGISTRATION_ID_NAME = 'X-IMS-EndRegistration-ID'
+    ACKNOWLEDGE_URL = 'VND-IMS-ACKNOWLEDGE-URL'
+
+    HTTP_ACKNOWLEDGE_URL = 'HTTP_VND_IMS_ACKNOWLEDGE_URL'
 
     def create_tool_proxy tool_consumer_profile, tool_proxy_guid, disposition
       tool_provider_registry = Rails.application.config.tool_provider_registry
       tool_proxy = {}
-      tool_proxy['@context'] = [
-        'http://purl.imsglobal.org/ctx/lti/v2/ToolProxy'
-      ]
+      # clone from provided TCP
+      tool_proxy['@context'] = tool_consumer_profile['@context'].clone
       tool_proxy['@type'] = 'ToolProxy'
       tool_proxy['@id'] = "ToolProxyProposal_at_#{Time.now.utc.iso8601}"
 
       tool_proxy['lti_version'] = 'LTI-2p0'
-      tool_proxy['tool_proxy_guid'] = tool_proxy_guid
-      # Fails conformance since it's undefined and checking is too strict
-      tool_proxy['disposition'] = disposition
+      tool_proxy['tool_proxy_guid'] = tool_proxy_guid if disposition == 'reregister'
+
       tool_proxy['tool_consumer_profile'] = self.tc_profile_url
       tool_proxy['tool_profile'] = JSON.load( tool_profile_json )
       tool_proxy['security_contract'] = resolve_security_contract( tool_consumer_profile )
@@ -59,22 +59,23 @@ module Lti2Tp
           status = create_status(false, nil, "#{err_code}-#{err_msg}")
           return status
         end
-        if disposition == 'register'
-        # get guid from the response returned by the TC
-        tool_proxy['tool_proxy_guid'] = tool_proxy_response['tool_proxy_guid']
+        if disposition.blank? || disposition == 'register' 
+          # get guid from the response returned by the TC
+          tool_proxy['tool_proxy_guid'] = tool_proxy_response['tool_proxy_guid']
 
-        # substitute tool_proxy_guid now in the Proxy where needed
-        tool_proxy_wrapper = JsonWrapper.new( tool_proxy )
-        tool_proxy_wrapper.substitute_text_in_all_nodes( '{', '}', { 'tool_proxy_guid' => tool_proxy['tool_proxy_guid'] } )
 
-        self.tool_proxy_json = tool_proxy.to_json
+          # substitute tool_proxy_guid now in the Proxy where needed
+          tool_proxy_wrapper = JsonWrapper.new( tool_proxy )
+          tool_proxy_wrapper.substitute_text_in_all_nodes( '{', '}', { 'tool_proxy_guid' => tool_proxy['tool_proxy_guid'] } )
+
+          self.tool_proxy_json = tool_proxy.to_json
           self.status = disposition
-        self.save!
+          self.save!
 
           status = create_status(true, tool_proxy_wrapper.first_at('tool_proxy_guid'))
-      else
+        else
           status = create_status(true)
-      end
+        end
       else
         status = create_status(false, nil, "Can't access ToolProxy")
       end
@@ -138,8 +139,11 @@ module Lti2Tp
 
       puts "Register request: #{signed_request.signature_base_string}"
       puts "Register secret: #{self.reg_password}"
+
+      headers = {}
+      (headers[ACKNOWLEDGE_URL] = self.end_registration_id) if disposition == 'reregister'
       response = invoke_service(signed_request, Rails.application.config.wire_log, "#{label} ToolProxy with ToolConsumer",
-          END_REGISTRATION_ID_NAME => self.end_registration_id)
+          headers)
       if response.code.between?( 200, 202 )
         response_body = response.body
         response_content = JSON.load( response_body ) unless response_body.strip.empty?
@@ -160,7 +164,7 @@ module Lti2Tp
       services_offered.each { |service_offered|
         tool_service = {}
         tool_service['@type'] = 'RestServiceProfile'
-        tool_service['service'] = service_offered['endpoint']
+        tool_service['service'] = service_offered['@id']
         tool_service['action'] = service_offered['action']
 
         # following may be wrong...check on proper construction of RestService
