@@ -4,6 +4,7 @@ Lti2Tp::Engine.routes.draw do
 
   match 'reregistrations' => 'registrations#reregister', :via => :post
 
-  match 'tool_proxies' => 'registrations#end_registration', :via => :post
+  match 'tool_proxies/:id' => 'registrations#complete_reregistration', :via => :put
+  match 'tool_proxies/:id' => 'registrations#complete_reregistration', :via => :delete
 
 end
