@@ -83,11 +83,11 @@ ActiveAdmin.register_page 'Course Page' do
       }
 
       # prepare resource key and list
-      resources = Lti2Tc::Resource.order('name').all
+      resources = Lti2Tc::Resource.order('resource_name').all
       resource_list = []
       resource_map = {}
       resources.each do |r|
-        resource_label = "#{r.name}: #{r.description}"
+        resource_label = "#{r.resource_name}: #{r.description}"
         resource_list << r.id
         resource_map[r.id] = resource_label
       end
